Add animation for username on welcome page
